modularise the package: no code changes, everything just distributed into modules
